requirements.txt
absl-py==2.1.0
alembic==1.13.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
category-encoders==2.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.0.1
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
icecream==2.1.3
idna==3.6
importlib-metadata==5.2.0
importlib-resources==6.0.0
ipykernel==6.29.3
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mlflow==1.30.0
mrmr-selection==0.2.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.25.1
oauthlib==3.2.2
opt-einsum==3.3.0
optuna==3.5.0
overrides==7.7.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
polars==0.20.23
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.1.1
scipy==1.12.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.46
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.2
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-addons==0.23.0
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typeguard==2.13.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.16
waitress==2.1.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.17.0
